requirements.txt
streamlit
numpy
pillow
streamlit
numpy
pillow
tensorflow
opencv-python
